+++ src/crypto/webcrypto.js
@@ -1,7 +1,11 @@
 'use strict'
 
 module.exports = function getWebCrypto () {
-  const WebCrypto = require('node-webcrypto-ossl')
-  const webCrypto = new WebCrypto()
-  return webCrypto
+  try {
+    const WebCrypto = require('node-webcrypto-ossl')
+    const webCrypto = new WebCrypto()
+    return webCrypto
+  } catch (err) {
+    // fallback to other things
+  }
 }
